# Speedup DistinctString and ScopedDistinctString collectors