--- a/.gitignore
+++ b/.gitignore
@@ -3,6 +3,7 @@
 .idea
 .vscode
 *.test
+*.pprof
 /bin
 /cmd/tempo-cli/tempo-cli
 /cmd/tempo-query/tempo-query