--- a/pkg/collector/distinct_string_collector_test.go
+++ b/pkg/collector/distinct_string_collector_test.go
@@ -2,6 +2,7 @@ package collector
 
 import (
 	"fmt"
+	"strconv"
 	"sync"
 	"testing"
 
@@ -52,4 +53,55 @@ func TestDistinctStringCollectorIsSafe(t *testing.T) {
 
 	require.Equal(t, len(d.Strings()), 10*100)
 	require.False(t, d.Exceeded())
+}
+
+func BenchmarkDistinctStringCollect(b *testing.B) {
+	// simulate 100 ingesters, each returning 10_000 tag values
+	numIngesters := 100
+	numTagValuesPerIngester := 10_000
+	ingesterStrings := make([][]string, numIngesters)
+	for i := 0; i < numIngesters; i++ {
+		strings := make([]string, numTagValuesPerIngester)
+		for j := 0; j < numTagValuesPerIngester; j++ {
+			strings[j] = fmt.Sprintf("string_%d_%d", i, j)
+		}
+		ingesterStrings[i] = strings
+	}
+
+	limits := []int{
+		0,          // no limit
+		100_000,    // 100KB
+		1_000_000,  // 1MB
+		10_000_000, // 10MB
+	}
+
+	b.ResetTimer() // to exclude the setup time for generating tag values
+	for _, lim := range limits {
+		b.Run("uniques_limit:"+strconv.Itoa(lim), func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				distinctStrings := NewDistinctString(lim)
+				for _, values := range ingesterStrings {
+					for _, v := range values {
+						if distinctStrings.Collect(v) {
+							break // stop early if limit is reached
+						}
+					}
+				}
+			}
+		})
+
+		b.Run("duplicates_limit:"+strconv.Itoa(lim), func(b *testing.B) {
+			for n := 0; n < b.N; n++ {
+				distinctStrings := NewDistinctString(lim)
+				for i := 0; i < numIngesters; i++ {
+					for j := 0; j < numTagValuesPerIngester; j++ {
+						// collect first item to simulate duplicates
+						if distinctStrings.Collect(ingesterStrings[i][0]) {
+							break // stop early if limit is reached
+						}
+					}
+				}
+			}
+		})
+	}
 }