--- a/pkg/collector/scoped_distinct_string.go
+++ b/pkg/collector/scoped_distinct_string.go
@@ -1,6 +1,8 @@
 package collector
 
-import "sync"
+import (
+	"sync"
+)
 
 type ScopedDistinctString struct {
 	cols        map[string]*DistinctString