… CHANGELOG.md
… pkg/collector/scoped_distinct_string_test.go
… tempodb/tempodb_search_test.go